improved the installer module to support balthazar's CD

+++ src/install.cpp
@@ -24,8 +24,6 @@ void install_TA_files( String def_path )
 									"/media/dvdrecorder/", "/media/dvdrecorder0/", "/media/dvdrecorder1/", "/media/dvdrecorder2/", "/media/dvdrecorder3/", "/media/dvdrecorder4/", "/media/dvdrecorder5/" };
 #endif
 
-	allegro_message( "please mount/insert your TA cdrom now (TA CD 1)" );
-
 	String path_to_TA_cd = "";
 
 	if( def_path != "" ) {						// Explicit path given
@@ -39,9 +37,31 @@ void install_TA_files( String def_path )
 	else									// Look for a path where we can find totala3.hpi
 		for(int i = 0 ; i < nb_possible_path ; i++ ) {
 			path_to_TA_cd = possible_path[ i ];
-			if( exists( ( path_to_TA_cd + "totala3.hpi" ).c_str() )
+			if( ( exists( ( path_to_TA_cd + "totala3.hpi" ).c_str() ) || exists( ( path_to_TA_cd + "totala4.hpi" ).c_str() ) )
 			 && exists( ( path_to_TA_cd + "totala2.hpi" ).c_str() ) )		break;		// We found a path to TA's cd ( in fact to needed files )
 			}
+
+	if( !( exists( ( path_to_TA_cd + "totala3.hpi" ).c_str() ) || exists( ( path_to_TA_cd + "totala4.hpi" ).c_str() ) )
+	 || !exists( ( path_to_TA_cd + "totala2.hpi" ).c_str() ) ) {															// We need TA's cd ( in fact only the files )
+		allegro_message( "please mount/insert your TA cdrom now (TA CD 1)" );
+
+		path_to_TA_cd = "";
+
+		if( def_path != "" ) {						// Explicit path given
+			path_to_TA_cd = def_path;
+			for( int i = 0 ; i < path_to_TA_cd.size() ; i++ )			// Use UNIX like path
+				if( path_to_TA_cd[ i ] == '\\' )
+					path_to_TA_cd[ i ] = '/';
+			if( path_to_TA_cd[ path_to_TA_cd.size() ] != '/' )			// Check it ends with a '\'
+				path_to_TA_cd += "/";
+			}
+		else									// Look for a path where we can find totala3.hpi
+			for(int i = 0 ; i < nb_possible_path ; i++ ) {
+				path_to_TA_cd = possible_path[ i ];
+				if( ( exists( ( path_to_TA_cd + "totala3.hpi" ).c_str() ) || exists( ( path_to_TA_cd + "totala4.hpi" ).c_str() ) )
+				 && exists( ( path_to_TA_cd + "totala2.hpi" ).c_str() ) )		break;		// We found a path to TA's cd ( in fact to needed files )
+				}
+		}
 
 	HPIManager=new cHPIHandler( path_to_TA_cd );
 
@@ -127,12 +147,15 @@ void install_TA_files( String def_path )
 		success = false;
 		
 	if( success ) {
-		allegro_message( "please mount/insert your TA cdrom now (TA CD 2) if you want to install campaign files" );
+	
+		if( !exists( ( path_to_TA_cd + "totala4.hpi" ).c_str() ) ) {
+			allegro_message( "please mount/insert your TA cdrom now (TA CD 2) if you want to install campaign files" );
 
-										// Look for a path where we can find totala4.hpi
-		for(int i = 0 ; i < nb_possible_path ; i++ ) {
-			path_to_TA_cd = possible_path[ i ];
-			if( exists( ( path_to_TA_cd + "totala4.hpi" ).c_str() ) )		break;		// We found a path to TA's cd ( in fact to needed files )
+											// Look for a path where we can find totala4.hpi
+			for(int i = 0 ; i < nb_possible_path ; i++ ) {
+				path_to_TA_cd = possible_path[ i ];
+				if( exists( ( path_to_TA_cd + "totala4.hpi" ).c_str() ) )		break;		// We found a path to TA's cd ( in fact to needed files )
+				}
 			}
 
 		if( exists( ( path_to_TA_cd + "totala4.hpi" ).c_str() ) ) {
@@ -169,6 +192,69 @@ void install_TA_files( String def_path )
 				}
 			}
 
+		{
+			if( !exists( ( path_to_TA_cd + "cc/ccdata.ccx" ).c_str() ) && !exists( ( path_to_TA_cd + "install/ccdata.ccx" ).c_str() ) )
+				allegro_message( "please mount/insert your TA:Core Contingency cdrom now (TA:CC CD) if you have it" );
+			bool cc_success = true;
+			int nb_files = 3;
+			int Y = 0;
+			String file_to_copy[] = { "cc/ccdata.ccx", "cc/ccmaps.ccx", "cc/ccmiss.ccx" };
+			String file_to_copy_alternative[] = { "install/ccdata.ccx", "install/ccmaps.ccx", "install/ccmiss.ccx" };
+			String file_destination[] = { "ccdata.ccx", "ccmaps.ccx", "ccmiss.ccx" };
+			rectfill( screen, 0, 0, 640, 480, 0x7F7F7F );
+											// Look for a path where we can find the files we need
+			for(int i = 0 ; i < nb_possible_path ; i++ ) {
+				path_to_TA_cd = possible_path[ i ];
+				if( exists( ( path_to_TA_cd + file_to_copy[ 0 ] ).c_str() ) )		break;		// We found a path to TA's cd ( in fact to needed files )
+				if( exists( ( path_to_TA_cd + file_to_copy_alternative[ 0 ] ).c_str() ) )		break;		// We found a path to TA's cd ( in fact to needed files )
+				}
+			if( exists( ( path_to_TA_cd + file_to_copy_alternative[ 0 ] ).c_str() ) )			// We found a path to an alternative TA CD
+				for( int i = 0 ; i < nb_files ; i++ )
+					file_to_copy[i] = file_to_copy_alternative[i];
+			for( int i = 0 ; i < nb_files ; i++ )
+				if( exists( ( path_to_TA_cd + file_to_copy[ i ] ).c_str() ) && cc_success ) {
+					textprintf_centre_ex( screen, font, 320, 40 + Y, 0x0, -1, format( "Copying %s (step %d/%d)", file_to_copy[ i ].c_str(), 1+i, nb_files ).c_str() );
+					rectfill( screen, 100, 60 + Y, 540, 80 + Y, makecol( 255, 0, 0 ) );
+					textprintf_centre_ex( screen, font, 320, 66 + Y, 0x0, -1, "0%%" );
+
+					FILE *src = TA3D_OpenFile( path_to_TA_cd + file_to_copy[ i ], "rb" );
+					if( src ) {
+						FILE *dst = TA3D_OpenFile( file_destination[ i ], "wb" );
+						if( dst ) {
+							int limit = FILE_SIZE( ( path_to_TA_cd + file_to_copy[ i ] ).c_str() );
+							byte *buf = new byte[ buf_size ];			// a 1Mo buffer
+
+							for( int pos = 0 ; pos < limit ; pos+= buf_size ) {
+								int read_size = min( buf_size, limit-pos );
+								fread( buf, read_size, 1, src );
+								fwrite( buf, read_size, 1, dst );
+
+								rectfill( screen, 100, 60 + Y, 540, 80 + Y, makecol( 255, 0, 0 ) );
+								rectfill( screen, 100, 60 + Y, 100 + 440 * (pos+read_size>>10) / (limit>>10), 80 + Y, makecol( 255, 255, 0 ) );
+								textprintf_centre_ex( screen, font, 320, 66 + Y, 0x0, -1, "%d%%", 100 * (pos+read_size>>10) / (limit>>10) );
+								}
+
+							delete buf;
+
+							fflush( dst );
+
+							fclose( dst );
+							}
+						else
+							cc_success = false;
+						fclose( src );
+						}
+					else
+						cc_success = false;
+					Y += 100;
+					if( Y + 80 >= 440 && i + 1 < nb_files ) {
+						Y -= 100;
+						blit( screen, screen, 0, 100, 0, 0, SCREEN_W, SCREEN_H - 100 );
+						}
+					}
+				else
+					cc_success = false;
+		}
 		{
 			allegro_message( "please mount/insert your TA:Battle Tactics cdrom now (TA:BT CD) if you have it" );
 			bool bt_success = true;
@@ -225,64 +311,6 @@ void install_TA_files( String def_path )
 					}
 				else
 					bt_success = false;
-		}
-
-		{
-			allegro_message( "please mount/insert your TA:Core Contingency cdrom now (TA:CC CD) if you have it" );
-			bool cc_success = true;
-			int nb_files = 3;
-			int Y = 0;
-			String file_to_copy[] = { "cc/ccdata.ccx", "cc/ccmaps.ccx", "cc/ccmiss.ccx" };
-			String file_destination[] = { "ccdata.ccx", "ccmaps.ccx", "ccmiss.ccx" };
-			rectfill( screen, 0, 0, 640, 480, 0x7F7F7F );
-											// Look for a path where we can find the files we need
-			for(int i = 0 ; i < nb_possible_path ; i++ ) {
-				path_to_TA_cd = possible_path[ i ];
-				if( exists( ( path_to_TA_cd + file_to_copy[ 0 ] ).c_str() ) )		break;		// We found a path to TA's cd ( in fact to needed files )
-				}
-			for( int i = 0 ; i < nb_files ; i++ )
-				if( exists( ( path_to_TA_cd + file_to_copy[ i ] ).c_str() ) && cc_success ) {
-					textprintf_centre_ex( screen, font, 320, 40 + Y, 0x0, -1, format( "Copying %s (step %d/%d)", file_to_copy[ i ].c_str(), 1+i, nb_files ).c_str() );
-					rectfill( screen, 100, 60 + Y, 540, 80 + Y, makecol( 255, 0, 0 ) );
-					textprintf_centre_ex( screen, font, 320, 66 + Y, 0x0, -1, "0%%" );
-
-					FILE *src = TA3D_OpenFile( path_to_TA_cd + file_to_copy[ i ], "rb" );
-					if( src ) {
-						FILE *dst = TA3D_OpenFile( file_destination[ i ], "wb" );
-						if( dst ) {
-							int limit = FILE_SIZE( ( path_to_TA_cd + file_to_copy[ i ] ).c_str() );
-							byte *buf = new byte[ buf_size ];			// a 1Mo buffer
-
-							for( int pos = 0 ; pos < limit ; pos+= buf_size ) {
-								int read_size = min( buf_size, limit-pos );
-								fread( buf, read_size, 1, src );
-								fwrite( buf, read_size, 1, dst );
-
-								rectfill( screen, 100, 60 + Y, 540, 80 + Y, makecol( 255, 0, 0 ) );
-								rectfill( screen, 100, 60 + Y, 100 + 440 * (pos+read_size>>10) / (limit>>10), 80 + Y, makecol( 255, 255, 0 ) );
-								textprintf_centre_ex( screen, font, 320, 66 + Y, 0x0, -1, "%d%%", 100 * (pos+read_size>>10) / (limit>>10) );
-								}
-
-							delete buf;
-
-							fflush( dst );
-
-							fclose( dst );
-							}
-						else
-							cc_success = false;
-						fclose( src );
-						}
-					else
-						cc_success = false;
-					Y += 100;
-					if( Y + 80 >= 440 && i + 1 < nb_files ) {
-						Y -= 100;
-						blit( screen, screen, 0, 100, 0, 0, SCREEN_W, SCREEN_H - 100 );
-						}
-					}
-				else
-					cc_success = false;
 		}
 		}
